requirements.txt
# ISAR Image Analysis - Deep Learning Product
# Core ML/DL Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# GUI Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Report Generation
reportlab>=4.0.0
jinja2>=3.1.0
fpdf2>=2.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Model Interpretability
grad-cam>=1.4.8
captum>=0.6.0

# Logging and Monitoring
tensorboard>=2.14.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
